+++ C++/src/ACCoRD/DensityGridMovingPolysEst.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 United States Government as represented by
+ * Copyright (c) 2016-2017 United States Government as represented by
  * the National Aeronautics and Space Administration.  No copyright
  * is claimed in the United States under Title 17, U.S.Code. All Other
  * Rights Reserved.
@@ -26,12 +26,12 @@ namespace larcfm {
 
 DensityGridMovingPolysEst::DensityGridMovingPolysEst(const Plan& p, int buffer, double squareSize, double gs_, const std::vector<PolyPath>& ps, const std::vector<PolyPath>& containment) :
 		DensityGridTimed(p, buffer, squareSize) {
-	for (int i = 0; i < ps.size(); i++) {
+	for (int i = 0; i < (int) ps.size(); i++) {
 		Triple<Plan,double,double> pp = ps[i].buildPlan();
 		plans.push_back(pp);
 	}
 	if (containment.size() > 0) {
-		for (int i = 0; i < containment.size(); i++) {
+		for (int i = 0; i < (int) containment.size(); i++) {
 			// static plans have already been handled in StratwayCore with a call to setWeightsWithin()
 			if (!containment[i].isStatic()) {
 				Triple<Plan,double,double> cc = containment[i].buildPlan();
@@ -57,7 +57,7 @@ double DensityGridMovingPolysEst::getWeightT(int x, int y, double t) const {
 	double cost = 0;
 	Position cent = center(x,y);
 	// disallow anything within one of the weather cells
-	for (int i = 0; i < plans.size(); i++) {
+	for (int i = 0; i < (int) plans.size(); i++) {
 		if (t >= plans[i].first.getFirstTime() && t <= plans[i].first.getLastTime()) {
 			Position p = plans[i].first.position(t);
 			double D = plans[i].second;
@@ -71,7 +71,7 @@ double DensityGridMovingPolysEst::getWeightT(int x, int y, double t) const {
 
 	// disallow anything outside of one of the contains
 	bool within = (contains.size() == 0);
-	for (int i = 0; i < contains.size(); i++) {
+	for (int i = 0; i < (int) contains.size(); i++) {
 		if (t >= contains[i].first.getFirstTime() && t <= contains[i].first.getLastTime()) {
 			Position p = contains[i].first.position(t);
 			double D = contains[i].second;
